chore: update save context on reorder issue
