+++ packages/hoppscotch-common/src/helpers/rest/tab.ts
@@ -1,6 +1,6 @@
 import { v4 as uuidV4 } from "uuid"
 import { isEqual } from "lodash-es"
-import { reactive, watch, computed, ref, shallowReadonly } from "vue"
+import { reactive, watch, computed, ref, shallowReadonly, Ref } from "vue"
 import { HoppRESTDocument, HoppRESTSaveContext } from "./document"
 import { refWithControl } from "@vueuse/core"
 import { HoppRESTResponse } from "../types/HoppRESTResponse"
@@ -192,45 +192,8 @@ export function getTabRefWithSaveContext(ctx: HoppRESTSaveContext) {
   return null
 }
 
-export function resolveSaveContextOnReorder(
-  lastIndex: number,
-  newIndex: number,
-  folderPath: string
-) {
-  if (newIndex > lastIndex) newIndex-- // there is a issue when going down? better way to resolve this?
-  if (lastIndex === newIndex) return
-
-  const effectedIndexes = getAffectedIndexes(lastIndex, newIndex)
-
-  getActiveTabs().value.forEach((tab) => {
-    if (
-      tab.document.saveContext?.originLocation === "user-collection" &&
-      tab.document.saveContext.folderPath === folderPath &&
-      effectedIndexes.has(tab.document.saveContext.requestIndex)
-    ) {
-      const newIndex = effectedIndexes.get(
-        tab.document.saveContext.requestIndex
-      )!
-      tab.document.saveContext.requestIndex = newIndex
-    }
-  })
-
-  console.log("actually moving", {
-    getActiveTabs: getActiveTabs().value,
-  })
-}
-
-function getAffectedIndexes(oldIndex: number, newIndex: number) {
-  const indexes = new Map<number, number>()
-  indexes.set(oldIndex, newIndex)
-  if (oldIndex < newIndex) {
-    for (let i = oldIndex + 1; i <= newIndex; i++) {
-      indexes.set(i, i - 1)
-    }
-  } else {
-    for (let i = oldIndex - 1; i >= newIndex; i--) {
-      indexes.set(i, i + 1)
-    }
-  }
-  return indexes
+export function getTabsRefTo(func: (tab: HoppRESTTab) => boolean) {
+  return Array.from(tabMap.values())
+    .filter(func)
+    .map((tab) => getTabRef(tab.id))
 }
